+++ src/test/compile-fail/borrowck-closures-use-after-free.rs
@@ -12,9 +12,10 @@
 // cannot also be supplied a borrowed version of that
 // variable's contents. Issue #11192.
 
+#![feature(box_syntax)]
 
 struct Foo {
-  x: int
+  x: isize
 }
 
 impl Drop for Foo {
