+++ src/test/compile-fail/borrowck-issue-14498.rs
@@ -11,11 +11,13 @@
 // This tests that we can't modify Box<&mut T> contents while they
 // are borrowed.
 
-struct A { a: int }
-struct B<'a> { a: Box<&'a mut int> }
+#![feature(box_syntax)]
+
+struct A { a: isize }
+struct B<'a> { a: Box<&'a mut isize> }
 
 fn borrow_in_var_from_var() {
-    let mut x: int = 1;
+    let mut x: isize = 1;
     let y = box &mut x;
     let p = &y;
     let q = &***p;
@@ -35,7 +37,7 @@ fn borrow_in_var_from_field() {
 }
 
 fn borrow_in_field_from_var() {
-    let mut x: int = 1;
+    let mut x: isize = 1;
     let y = B { a: box &mut x };
     let p = &y.a;
     let q = &***p;
