+++ src/test/compile-fail/borrowck-move-subcomponent.rs
@@ -11,9 +11,10 @@
 // Tests that the borrow checker checks all components of a path when moving
 // out.
 
+#![feature(box_syntax)]
 
 struct S {
-  x : Box<int>
+  x : Box<isize>
 }
 
 fn f<T>(_: T) {}
