+++ src/test/compile-fail/coherence-blanket-conflicts-with-blanket-implemented.rs
@@ -16,23 +16,23 @@ use std::default::Default;
 // for the same type (though this crate doesn't).
 
 trait MyTrait {
-    fn get(&self) -> uint;
+    fn get(&self) -> usize;
 }
 
 trait Even { }
 
 trait Odd { }
 
-impl Even for int { }
+impl Even for isize { }
 
-impl Odd for uint { }
+impl Odd for usize { }
 
 impl<T:Even> MyTrait for T { //~ ERROR E0119
-    fn get(&self) -> uint { 0 }
+    fn get(&self) -> usize { 0 }
 }
 
 impl<T:Odd> MyTrait for T {
-    fn get(&self) -> uint { 0 }
+    fn get(&self) -> usize { 0 }
 }
 
 fn main() { }
