+++ src/test/compile-fail/dst-rvalue.rs
@@ -10,13 +10,15 @@
 
 // Check that dynamically sized rvalues are forbidden
 
+#![feature(box_syntax)]
+
 pub fn main() {
     let _x: Box<str> = box *"hello world";
     //~^ ERROR E0161
-    //~^^ ERROR cannot move out of dereference
+    //~^^ ERROR cannot move out of borrowed content
 
-    let array: &[int] = &[1, 2, 3];
-    let _x: Box<[int]> = box *array;
+    let array: &[isize] = &[1, 2, 3];
+    let _x: Box<[isize]> = box *array;
     //~^ ERROR E0161
-    //~^^ ERROR cannot move out of dereference
+    //~^^ ERROR cannot move out of borrowed content
 }
