+++ src/test/compile-fail/enum-in-scope.rs
@@ -8,7 +8,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-struct hello(int);
+struct hello(isize);
 
 fn main() {
     let hello = 0; //~ERROR declaration of `hello` shadows
