+++ src/test/compile-fail/issue-17718-const-naming.rs
@@ -10,7 +10,7 @@
 
 #[deny(warnings)]
 
-const foo: int = 3;
+const foo: isize = 3;
 //~^ ERROR: should have an uppercase name such as
 //~^^ ERROR: constant item is never used
