+++ src/test/compile-fail/issue-4935.rs
@@ -10,5 +10,5 @@
 
 // Regression test for issue #4935
 
-fn foo(a: uint) {}
+fn foo(a: usize) {}
 fn main() { foo(5, 6) } //~ ERROR this function takes 1 parameter but 2 parameters were supplied
