+++ src/test/compile-fail/issue-5439.rs
@@ -8,17 +8,18 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+#![feature(box_syntax)]
 
 struct Foo {
-    foo: int,
+    foo: isize,
 }
 
 struct Bar {
-    bar: int,
+    bar: isize,
 }
 
 impl Bar {
-    fn make_foo (&self, i: int) -> Box<Foo> {
+    fn make_foo (&self, i: isize) -> Box<Foo> {
         return box Foo { nonexistent: self, foo: i }; //~ ERROR: no field named
     }
 }
