+++ src/test/compile-fail/issue-6801.rs
@@ -12,17 +12,18 @@
 // transferring ownership of the owned box before invoking the stack
 // closure results in a crash.
 
+#![feature(box_syntax)]
 
-fn twice(x: Box<uint>) -> uint {
+fn twice(x: Box<usize>) -> usize {
      *x * 2
 }
 
-fn invoke<F>(f: F) where F: FnOnce() -> uint {
+fn invoke<F>(f: F) where F: FnOnce() -> usize {
      f();
 }
 
 fn main() {
-      let x  : Box<uint>  = box 9;
+      let x  : Box<usize>  = box 9;
       let sq =  |:| { *x * *x };
 
       twice(x); //~ ERROR: cannot move out of
