+++ src/test/compile-fail/issue-7364.rs
@@ -8,11 +8,12 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+#![feature(box_syntax)]
 
 use std::cell::RefCell;
 
 // Regresion test for issue 7364
-static boxed: Box<RefCell<int>> = box RefCell::new(0);
+static boxed: Box<RefCell<isize>> = box RefCell::new(0);
 //~^ ERROR statics are not allowed to have custom pointers
 //~| ERROR: the trait `core::marker::Sync` is not implemented for the type
 //~| ERROR: the trait `core::marker::Sync` is not implemented for the type
