+++ src/test/compile-fail/lint-owned-heap-memory.rs
@@ -10,10 +10,10 @@
 
 #![allow(dead_code)]
 #![forbid(box_pointers)]
-
+#![feature(box_syntax)]
 
 struct Foo {
-    x: Box<int> //~ ERROR type uses owned
+    x: Box<isize> //~ ERROR type uses owned
 }
 
 fn main() {
